# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  order.CreateOrderRequest:
    properties:
      customer_id:
        type: string
      items:
        description: gt=0 slice validation
        items:
          $ref: '#/definitions/order.OrderItemRequest'
        type: array
    required:
    - customer_id
    - items
    type: object
  order.OrderItemRequest:
    properties:
      product_id:
        type: string
      quantity:
        type: integer
      unit_price:
        type: number
    required:
    - product_id
    - quantity
    - unit_price
    type: object
  order.OrderItemResponse:
    properties:
      category_name:
        type: string
      id:
        type: string
      product_id:
        type: string
      product_name:
        type: string
      quantity:
        type: integer
      unit_price:
        type: number
    type: object
  order.OrderResponse:
    properties:
      created_at:
        type: string
      customer_email:
        type: string
      customer_id:
        type: string
      customer_name:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/order.OrderItemResponse'
        type: array
      total_price:
        type: number
      total_quantity:
        type: integer
    type: object
  product.CategoryResponse:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  product.CreateProductRequest:
    properties:
      category_id:
        type: string
      description:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      price:
        type: number
      stock_quantity:
        minimum: 0
        type: integer
    required:
    - category_id
    - name
    - price
    type: object
  product.ProductResponse:
    properties:
      category:
        $ref: '#/definitions/product.CategoryResponse'
      description:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      price:
        type: number
      stock_quantity:
        type: integer
      total_sold:
        type: integer
    type: object
  product.UpdateProductRequest:
    properties:
      category_id:
        type: string
      description:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      price:
        type: number
      stock_quantity:
        minimum: 0
        type: integer
    required:
    - category_id
    - name
    - price
    type: object
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API Server for Order and Dashboard Management.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Assignment PTES API
  version: "1.0"
paths:
...
  /orders:
    get:
      description: Retrieve a paginated list of orders with basic customer info
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/order.OrderResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Place a new order with multiple items. Calculates total price and
        quantity automatically.
      parameters:
      - description: Order Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/order.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/order.OrderResponse'
        "400":
          description: Invalid input or empty items
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Customer or Product not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new order
      tags:
      - orders
  /orders/{id}:
    delete:
      description: Remove an order record and its associated items
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete an order
      tags:
      - orders
    get:
      description: Retrieve full order details including all item descriptions and
        category names
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/order.OrderResponse'
        "404":
          description: Order not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get order details
      tags:
      - orders
  /products:
    get:
      description: Get a list of products with advanced filters (price, stock, category)
      parameters:
      - in: query
        name: category
        type: string
      - in: query
        name: max_price
        type: number
      - in: query
        name: max_stock
        type: integer
      - in: query
        name: min_price
        type: number
      - in: query
        name: min_stock
        type: integer
      - in: query
        name: name
        type: string
      - in: query
        name: page
        type: integer
      - in: query
        name: page_size
        type: integer
      - in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/product.ProductResponse'
            type: array
      summary: List products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Create a new product with category association and stock details
      parameters:
      - description: Product Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/product.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/product.ProductResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new product
      tags:
      - products
  /products/{id}:
    delete:
      description: Remove a product record from the database
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete product
      tags:
      - products
    get:
      description: Retrieve product information including its category details
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/product.ProductResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get product detail
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Update product price, stock, or other details by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/product.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/product.ProductResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update product
      tags:
      - products
swagger: "2.0"